+++ tests/test_data_connect.py
@@ -285,3 +285,49 @@ class TestDataConnectService:
         err_msg = "Config must be of type firebase_admin.dataconnect.ConnectorConfig"
         with pytest.raises(ValueError, match=err_msg):
             self.service.get_client(None)
+
+
+class TestDataConnectServiceIntegration:
+
+    def setup_method(self):
+        self.cred = testutils.MockCredential()
+        self.app1 = firebase_admin.initialize_app(self.cred, name="integ_app1")
+        self.app2 = firebase_admin.initialize_app(self.cred, name="integ_app2")
+
+        self.config1 = BASE_CONFIG
+        self.config2 = dataconnect.ConnectorConfig(
+            service_id="service2", location="us-east4", connector="conn2"
+        )
+        self.config1_copy = dataconnect.ConnectorConfig(
+            service_id="starterproject", location="us-east4", connector="my_connector"
+        )
+
+    def teardown_method(self, method):
+        del method
+        testutils.cleanup_apps()
+
+    def test_overall_client_retrieval_and_caching(self):
+        client1a = dataconnect.client(self.config1, app=self.app1)
+        client1b = dataconnect.client(self.config1_copy, app=self.app1)
+        client2 = dataconnect.client(self.config2, app=self.app1)
+
+        assert isinstance(client1a, dataconnect.DataConnect)
+        assert client1a.app is self.app1
+        assert client1a.config is self.config1
+
+        # Same config
+        assert client1b is client1a
+
+        # Different config
+        assert isinstance(client2, dataconnect.DataConnect)
+        assert client2.app is self.app1
+        assert client2.config is self.config2
+        assert client2 is not client1a
+
+        # Different app
+        client1_app2 = dataconnect.client(self.config1, app=self.app2)
+
+        assert isinstance(client1_app2, dataconnect.DataConnect)
+        assert client1_app2.app is self.app2
+        assert client1_app2.config is self.config1
+        assert client1_app2 is not client1a
